tests moved to general test folder
