+++ src/tests/unit/pages/thanks-page/thanks-page.variables.js
@@ -1,0 +1,27 @@
+export const order = {
+  _id: '619eb12f3c53565320f384c6',
+  orderNumber: '1637789999938',
+  recipient: {
+    firstName: 'John',
+    lastName: 'Dou',
+    phoneNumber: '380934850648'
+  },
+  delivery: {
+    byCourier: false,
+    sentBy: 'NOVAPOST',
+    courierOffice: 'Test office',
+    city: '',
+    street: '',
+    house: ''
+  }
+};
+
+export const state = {
+  isLightTheme: true,
+  language: 1,
+  loading: false,
+  currency: 0,
+  order,
+  paidOrderLoading: false,
+  user: {}
+};
